Add game stats module
